requirements.txt
gunicorn
flask
python-dotenv
PyMySQL
certifi